00 UNICEF/09 HDX script/requirements.txt
appdirs==1.4.4
attrs==21.4.0
basicauth==0.4.1
beautifulsoup4==4.10.0
boto3==1.21.21
botocore==1.24.21
cattrs==1.10.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
ckanapi==4.6
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
decorator==5.1.1
defopt==6.3.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.18.1
email-validator==1.1.3
et-xmlfile==1.1.0
exchangerates==0.3.4
greenlet==1.1.2
hdx-python-api==5.5.9
hdx-python-country==3.1.3
hdx-python-utilities==3.1.7
html5lib==1.1
idna==3.3
ijson==3.1.4
inflect==5.4.0
jmespath==1.0.0
jsonlines==3.0.0
jsonpath-ng==1.5.3
libhxl==4.24.1
linear-tsv==1.1.0
loguru==0.6.0
lxml==4.8.0
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
num2words==0.5.10
openpyxl==3.0.9
ply==3.11
pockets==0.9.1
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==22.0.0
pyphonetics==0.5.3
python-dateutil==2.8.2
python-io-wrapper==0.2
python-slugify==6.1.1
quantulum3==0.7.10
ratelimit==2.2.1
requests==2.27.1
requests-cache==0.9.3
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
sphinxcontrib-napoleon==0.7
SQLAlchemy==1.4.32
tabulator==1.53.5
text-unidecode==1.3
typing-inspect==0.7.1
typing_extensions==4.1.1
unicodecsv==0.14.1
Unidecode==1.3.4
url-normalize==1.4.3
urllib3==1.26.9
webencodings==0.5.1
win32-setctime==1.1.0
xlrd==2.0.1
xlrd3==1.1.0
